InfoRequest.java updated
endpoints output surrounded with try catch, returning ResponseEntity<> instead of bare types;
service modified and surrounded with try catch;

+++ src/main/java/com/novo/services/InfoRequestServiceImpl.java
@@ -36,6 +36,9 @@ public class InfoRequestServiceImpl implements InfoRequestService {
     @Override
     public boolean deleteInfoRequest(int infoRequestId) {
         try {
+            if(infoRequestRepo.findById(infoRequestId).isEmpty()) {
+                throw new Exception("Richiesta non trovata.");
+            }
             infoRequestRepo.deleteById(infoRequestId);
             return true;
         }
